Move karma packages to `assets/src`
