+++ assets/src/karma-cuj-reporter/.eslintrc
