+++ assets/src/karma-cuj-reporter/index.cjs
