+++ assets/src/karma-failed-tests-reporter/.eslintrc
