+++ assets/src/karma-failed-tests-reporter/index.cjs
