+++ assets/src/karma-fixture/actPromise.js
