+++ assets/src/karma-fixture/componentStub.js
