+++ assets/src/karma-fixture/events.js
