+++ assets/src/karma-fixture/init.js
