+++ assets/src/karma-puppeteer-client/client.js
