+++ assets/src/karma-puppeteer-client/index.cjs
