+++ assets/src/karma-puppeteer-launcher/.eslintrc
