+++ assets/src/karma-puppeteer-launcher/mouseWithDnd.cjs
