+++ assets/src/karma-puppeteer-launcher/snapshot.cjs
